backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.340
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.26.2
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1